created opening scene

+++ src/game/gameplayStates/DolphinChamber.java
@@ -71,7 +71,7 @@ public class DolphinChamber extends GamePlayState {
 	
 	public void additionalEnter(GameContainer container, StateBasedGame stateManager) {
 		if(this.isEntered()==false && m_waterDown == false) {
-			Scene s = new Scene(this, m_player, new int[][] {{3,13}, {3,4},{3,13}});
+			Scene s = new Scene(this, m_player, new float[][] {{3,13}, {3,4},{3,13}});
 			//s.setPlayerInvisible(true);
 			s.setCamera(true);
 			s.playScene();
